Add DRM/GBM module
