+++ README.md
@@ -68,7 +68,7 @@ Layouts
 - Pane role assignment (C=video, A, B) is a permutation over the 3 slots and can be rotated/swapped at runtime via r/R/t.
 
 Planned TODOs
-- Refactor the monolithic compositor into separate DRM/GBM, mpv embed, and UI modules.
+- Refactor the monolithic compositor into separate DRM/GBM, mpv embed, and UI modules. [DRM/GBM split done]
 - Support a variable number of terminal panes rather than the fixed A/B pair.
 - Add a layout option with transparent terminal panes overlaying the video background.
 - The function find_monospace_font and surrounding FreeType initialization code appear in both src/osd.c and src/term_pane.c, leading to duplication and potential drift.
